add -pprofDir profiling to LSP
